Update wx.js
Overview
Affected versions of this package are vulnerable to Arbitrary Code Injection. The package does not validate user input for the reqPostMaterial function, thereby passing unsanitized contents of the file parameter to an exec call. This could potentially allow attackers to run arbitrary commands in the system.

Remediation
We handle user input within `execFile` function to safely pass to argument untrusted input.

References
031-JS-WXCHANGBA